chore: Restructure store to match standardised architecture
