+++ static/js/store/types/components/Category.d.ts
@@ -1,0 +1,6 @@
+type Category = {
+  display_name: string;
+  name: string;
+};
+
+export default Category;
