+++ static/js/store/types/index.d.ts
@@ -1,0 +1,4 @@
+import Category from "./components/Category";
+import Package from "./components/Package";
+
+export { Category, Package };
