+++ static/js/store/utils/getArchitectures.ts
@@ -1,0 +1,34 @@
+function getArchitectures() {
+  return [
+    {
+      name: "",
+      display_name: "All",
+    },
+    {
+      name: "amd64",
+      display_name: "AMD64",
+    },
+    {
+      name: "arm64",
+      display_name: "ARM64",
+    },
+    {
+      name: "armhf",
+      display_name: "ARMHF",
+    },
+    {
+      name: "i386",
+      display_name: "I386",
+    },
+    {
+      name: "ppc64el",
+      display_name: "PPC64EL",
+    },
+    {
+      name: "s390x",
+      display_name: "S390X",
+    },
+  ];
+}
+
+export default getArchitectures;
