+++ static/js/store/utils/getCategoryOrder.ts
@@ -1,0 +1,15 @@
+function getCategoryOrder() {
+  return [
+    "development",
+    "games",
+    "social",
+    "productivity",
+    "utilities",
+    "music-and-audio",
+    "art-and-design",
+    "photo-and-video",
+    "server-and-cloud",
+  ];
+}
+
+export default getCategoryOrder;
