+++ static/js/store/utils/index.ts
@@ -1,0 +1,4 @@
+import getArchitectures from "./getArchitectures";
+import getCategoryOrder from "./getCategoryOrder";
+
+export { getArchitectures, getCategoryOrder };
